+++ packages/veui/src/components/Transfer/Transfer.vue
@@ -5,8 +5,8 @@ import { isEqual, find, difference, includes, omit, uniq, remove, isString } fro
 import ui from '../../mixins/ui'
 import input from '../../mixins/input'
 
-function defaultFilter (type, keyword, option, datasource) {
-  return includes(option.label, keyword)
+function defaultFilter (type, keyword, item, datasource) {
+  return includes(item.label, keyword)
 }
 
 export default {
@@ -48,15 +48,13 @@ export default {
     },
     keys: {
       type: [String, Function],
-      default () {
-        return (source) => source.value
-      }
+      default: 'value'
     }
   },
   data () {
     return {
-      candidateOptions: [],
-      selectedOptions: [],
+      candidateItems: [],
+      selectedItems: [],
       rootAllCount: 0,
       rootPartCount: 0,
 
@@ -70,7 +68,7 @@ export default {
     },
     realKeys () {
       if (isString(this.keys)) {
-        return (source) => source[this.keys]
+        return source => source[this.keys]
       }
 
       return this.keys
@@ -78,55 +76,55 @@ export default {
   },
   watch: {
     datasource: {
-      handler (v, oldV) {
-        if (!isEqual(v, oldV)) {
-          let walk = (datasource, options) => {
+      handler (val, oldVal) {
+        if (!isEqual(val, oldVal)) {
+          let walk = (datasource, items) => {
             datasource.forEach((source, index) => {
-              let option = omit(source, 'children')
-              option.value = this.realKeys(source)
+              let item = omit(source, 'children')
+              item.value = this.realKeys(source)
               if (this.hasChild(source)) {
-                this.$set(option, 'children', [])
-                walk(source.children, option.children)
+                this.$set(item, 'children', [])
+                walk(source.children, item.children)
               }
-              this.$set(options, index, option)
+              this.$set(items, index, item)
             })
           }
-          walk(this.datasource, this.candidateOptions)
+          walk(this.datasource, this.candidateItems)
 
           this.correct()
-          this.setSelectedOptions(this.cloneSelectedOptions())
+          this.setSelectedItems(this.cloneSelectedItems())
         }
       },
       immediate: true
     },
-    selected (v, oldV) {
-      if (difference(v, oldV).length || difference(v, this.getSelectedValuesFromSelectedOptions()).length) {
+    selected (val, oldVal) {
+      if (difference(val, oldVal).length || difference(val, this.getSelectedValuesFromSelectedItems()).length) {
         this.correct()
-        this.setSelectedOptions(this.cloneSelectedOptions())
+        this.setSelectedItems(this.cloneSelectedItems())
       }
     }
   },
   created () {
     this.correct()
-    this.setSelectedOptions(this.cloneSelectedOptions())
+    this.setSelectedItems(this.cloneSelectedItems())
   },
   methods: {
     // 判断节点是否被选中：
     // 1、如果是叶子节点，直接根据 selected 属性判断。
     // 2、如果是非叶子节点，则该节点下所有子级节点都全部选择了，当前节点才算被选中了。
-    isSelected (option) {
-      return this.hasChild(option) ? option.allCount === option.children.length : option.selected
+    isSelected (item) {
+      return this.hasChild(item) ? item.allCount === item.children.length : item.selected
     },
 
-    // 取出 this.selectedOptions 中的 values 值，返回一个一维数组。
-    getSelectedValuesFromSelectedOptions () {
+    // 取出 this.selectedItems 中的 values 值，返回一个一维数组。
+    getSelectedValuesFromSelectedItems () {
       let values = []
-      let walk = (options = this.selectedOptions) => {
-        options.forEach(option => {
-          if (this.hasChild(option)) {
-            walk(option.children)
+      let walk = (items = this.selectedItems) => {
+        items.forEach(item => {
+          if (this.hasChild(item)) {
+            walk(item.children)
           } else {
-            values.push(option.value)
+            values.push(item.value)
           }
         })
       }
@@ -134,27 +132,27 @@ export default {
       return values
     },
     emitSelect () {
-      this.$emit('select', this.getSelectedValuesFromSelectedOptions())
+      this.$emit('select', this.getSelectedValuesFromSelectedItems())
     },
     selectAll () {
       if (!this.isSelectable) {
         return
       }
 
-      let walk = (options) => {
-        options.forEach(option => {
-          if (this.hasChild(option)) {
-            this.setOptionCount(option, option.children.length, 0)
-            walk(option.children)
+      let walk = (items) => {
+        items.forEach(item => {
+          if (this.hasChild(item)) {
+            this.setItemCount(item, item.children.length, 0)
+            walk(item.children)
           } else {
-            this.setLeafSelected(option, true)
+            this.setLeafSelected(item, true)
           }
 
-          this.$set(option, 'visuallySelected', true)
+          this.$set(item, 'visuallySelected', true)
         })
       }
-      walk(this.candidateOptions)
-      this.setSelectedOptions(this.cloneSelectedOptions())
+      walk(this.candidateItems)
+      this.setSelectedItems(this.cloneSelectedItems())
       this.emitSelect()
     },
     removeAll () {
@@ -162,58 +160,58 @@ export default {
         return
       }
 
-      let walk = (options) => {
-        options.forEach(option => {
-          if (this.hasChild(option)) {
-            this.setOptionCount(option, 0, 0)
-            walk(option.children)
+      let walk = (items) => {
+        items.forEach(item => {
+          if (this.hasChild(item)) {
+            this.setItemCount(item, 0, 0)
+            walk(item.children)
           } else {
-            this.setLeafSelected(option, false)
+            this.setLeafSelected(item, false)
           }
 
-          this.$set(option, 'visuallySelected', false)
+          this.$set(item, 'visuallySelected', false)
         })
       }
-      walk(this.candidateOptions)
-      this.setSelectedOptions([])
+      walk(this.candidateItems)
+      this.setSelectedItems([])
       this.emitSelect()
     },
 
     // 从左侧添加选中项
     //
     // 接收的参数为：
-    //  option ：当前被选中的节点
+    //  item ：当前被选中的节点
     //
-    // 注意： parents 来自于 this.candidateOptions ，内部的节点数据和 this.candidateOptions 中的节点数据是相等的。
+    // 注意： parents 来自于 this.candidateItems ，内部的节点数据和 this.candidateItems 中的节点数据是相等的。
     //
     // 执行步骤为：
-    // 1、以 this.candidateOptions 为主要标记目标。
-    // 2、如果当前选中节点 option 是叶子节点，则设置其被选中（ option.selected=true ）。
-    // 3、如果当前选中节点 option 是非叶子节点，则在 option 中标记“全选（ option.allCount=option.children.length, option.partCount=0 ）”。
+    // 1、以 this.candidateItems 为主要标记目标。
+    // 2、如果当前选中节点 item 是叶子节点，则设置其被选中（ item.selected=true ）。
+    // 3、如果当前选中节点 item 是非叶子节点，则在 item 中标记“全选（ item.allCount=item.children.length, item.partCount=0 ）”。
     // 4、依次标记祖先节点，刷新祖先节点中的 allCount 和 partCount 计数。
-    // 5、如果 option 是非叶子节点，则标记所有子孙节点为“全选（ allCount=children.length, partCount=0, selected=true ）”。
-    // 6、从之前选中的节点树中（ this.selectedOptions ）中解析出每个节点的状态（目前只有展开/收起状态 expanded ）。
-    // 7、从 this.candidateOptions 中剥离出选中的节点及其父节点，并与6步中解析出的状态进行合并，得到新的 this.selectedOptions 。
+    // 5、如果 item 是非叶子节点，则标记所有子孙节点为“全选（ allCount=children.length, partCount=0, selected=true ）”。
+    // 6、从之前选中的节点树中（ this.selectedItems ）中解析出每个节点的状态（目前只有展开/收起状态 expanded ）。
+    // 7、从 this.candidateItems 中剥离出选中的节点及其父节点，并与6步中解析出的状态进行合并，得到新的 this.selectedItems 。
     // 8、抛出 select 事件，带上一维的选中的节点的 value 值数组。
-    select (option) {
+    select (item) {
       if (!this.isSelectable) {
         return
       }
 
-      let chain = this.findChain(this.candidateOptions, option.value)
-      option = chain[chain.length - 1]
+      let chain = this.findChain(this.candidateItems, item.value)
+      item = chain[chain.length - 1]
       let parents = chain.slice(0, chain.length - 1).reverse()
 
-      if (this.hasChild(option)) {
-        this.setOptionCount(option, option.children.length, 0)
+      if (this.hasChild(item)) {
+        this.setItemCount(item, item.children.length, 0)
       } else {
-        this.setLeafSelected(option, true)
+        this.setLeafSelected(item, true)
       }
-      this.$set(option, 'visuallySelected', true)
+      this.$set(item, 'visuallySelected', true)
 
       this.markParentsChain(parents)
-      this.selectAllChildren(option, true)
-      this.setSelectedOptions(this.cloneSelectedOptions())
+      this.selectAllChildren(item, true)
+      this.setSelectedItems(this.cloneSelectedItems())
 
       this.emitSelect()
     },
@@ -221,46 +219,46 @@ export default {
     // 从右侧移除选中项。
     //
     // 接收的参数为：
-    //  option :待移除的节点
+    //  item :待移除的节点
     //
-    // 注意：parents 来自于 this.selectedOptions ，内部的节点数据和 this.candidateOptions 中的节点数据是完全不相等的。
+    // 注意：parents 来自于 this.selectedItems ，内部的节点数据和 this.candidateItems 中的节点数据是完全不相等的。
     //
     // 执行步骤为：
-    // 1、以 this.candidateOptions 为主要标记目标。
-    // 2、按照 parents 的最右侧开始遍历（即从树根开始），找到 parents 对应到 this.candidateOptions 中的 parents 祖先节点数组 candidateParents。
-    // 3、找到 option 对应到 this.candidateOptions 中的节点 candidateOption。
-    // 4、如果 candidateOption 没有子孙节点，就直接标记 candidateOption 上的选中状态（ selected ）为 false 。
-    // 5、如果 candidateOption 有子孙节点，则标记 candidateOption 上的“子级中全选的节点总数（ allCount ）”为 0，“子级中部分选择的节点总数（ partCount ）”为0。
+    // 1、以 this.candidateItems 为主要标记目标。
+    // 2、按照 parents 的最右侧开始遍历（即从树根开始），找到 parents 对应到 this.candidateItems 中的 parents 祖先节点数组 candidateParents。
+    // 3、找到 item 对应到 this.candidateItems 中的节点 candidateItem。
+    // 4、如果 candidateItem 没有子孙节点，就直接标记 candidateItem 上的选中状态（ selected ）为 false 。
+    // 5、如果 candidateItem 有子孙节点，则标记 candidateItem 上的“子级中全选的节点总数（ allCount ）”为 0，“子级中部分选择的节点总数（ partCount ）”为0。
     // 6、依次刷新 candidateParents 中的 allCount 和 partCount 标记。
-    // 7、如果 candidateOption 有子孙节点的话，则将其子孙节点全部设为未选中状态（ allCount=0 、 partCount=0 、 selected=false ）。
-    // 8、从之前选中的节点树中（ this.selectedOptions ）中解析出每个节点的状态（目前只有展开/收起状态 expanded ）。
-    // 9、从 this.candidateOptions 中剥离出选中的节点及其父节点，并与8步中解析出的状态进行合并，得到新的 this.selectedOptions 。
+    // 7、如果 candidateItem 有子孙节点的话，则将其子孙节点全部设为未选中状态（ allCount=0 、 partCount=0 、 selected=false ）。
+    // 8、从之前选中的节点树中（ this.selectedItems ）中解析出每个节点的状态（目前只有展开/收起状态 expanded ）。
+    // 9、从 this.candidateItems 中剥离出选中的节点及其父节点，并与8步中解析出的状态进行合并，得到新的 this.selectedItems 。
     // 10、抛出 select 事件，带上一维的选中的节点的 value 值数组。
-    remove (option) {
+    remove (item) {
       if (!this.isSelectable) {
         return
       }
 
-      // 先找到在 this.candidateOptions 里面对应的 option 和 parents 数组
-      let chain = this.findChain(this.candidateOptions, option.value)
-      option = chain[chain.length - 1]
+      // 先找到在 this.candidateItems 里面对应的 item 和 parents 数组
+      let chain = this.findChain(this.candidateItems, item.value)
+      item = chain[chain.length - 1]
       let parents = chain.slice(0, chain.length - 1).reverse()
 
-      if (this.hasChild(option)) {
-        this.setOptionCount(option, 0, 0)
+      if (this.hasChild(item)) {
+        this.setItemCount(item, 0, 0)
       } else {
-        this.setLeafSelected(option, false)
+        this.setLeafSelected(item, false)
       }
-      this.$set(option, 'visuallySelected', false)
+      this.$set(item, 'visuallySelected', false)
 
       this.markParentsChain(parents)
-      this.selectAllChildren(option, false)
+      this.selectAllChildren(item, false)
 
-      this.setSelectedOptions(this.cloneSelectedOptions())
+      this.setSelectedItems(this.cloneSelectedItems())
       this.emitSelect()
     },
-    setSelectedOptions (options) {
-      this.selectedOptions.splice(0, this.selectedOptions.length, ...options)
+    setSelectedItems (items) {
+      this.selectedItems.splice(0, this.selectedItems.length, ...items)
     },
     // 更新祖先节点中的选择标记（ selected 、 allCount 、 partCount ）
     markParentsChain (parents) {
@@ -278,79 +276,79 @@ export default {
             allCount += 1
           }
         })
-        this.setOptionCount(parent, allCount, partCount)
+        this.setItemCount(parent, allCount, partCount)
         this.$set(parent, 'visuallySelected', this.isSelected(parent))
       })
     },
     checkOwnProperty (obj, property) {
       return Object.prototype.hasOwnProperty.call(obj, property)
     },
-    setLeafSelected (option, selected) {
-      this.$set(option, 'selected', selected)
-      if (this.checkOwnProperty(option, 'allCount')) {
-        this.$set(option, 'allCount', undefined)
+    setLeafSelected (item, selected) {
+      this.$set(item, 'selected', selected)
+      if (this.checkOwnProperty(item, 'allCount')) {
+        this.$set(item, 'allCount', undefined)
       }
-      if (this.checkOwnProperty(option, 'partCount')) {
-        this.$set(option, 'partCount', undefined)
-      }
-    },
-    setOptionCount (option, allCount, partCount) {
-      this.$set(option, 'allCount', allCount)
-      this.$set(option, 'partCount', partCount)
-      if (this.checkOwnProperty(option, 'selected')) {
-        this.$set(option, 'selected', undefined)
+      if (this.checkOwnProperty(item, 'partCount')) {
+        this.$set(item, 'partCount', undefined)
       }
     },
-    hasChild (option) {
-      return option.children && option.children.length
+    setItemCount (item, allCount, partCount) {
+      this.$set(item, 'allCount', allCount)
+      this.$set(item, 'partCount', partCount)
+      if (this.checkOwnProperty(item, 'selected')) {
+        this.$set(item, 'selected', undefined)
+      }
     },
-    // 从 this.candidateOptions 中深克隆一份 selectedOptions 。
-    cloneSelectedOptions () {
-      let walk = (candidateOptions, selectedOptions) => {
-        let newSelectedOptions = []
-        candidateOptions.forEach(candidateOption => {
+    hasChild (item) {
+      return item.children && item.children.length
+    },
+    // 从 this.candidateItems 中深克隆一份 selectedItems 。
+    cloneSelectedItems () {
+      let walk = (candidateItems, selectedItems) => {
+        let newSelectedItems = []
+        candidateItems.forEach(candidateItem => {
           if (
-            candidateOption.allCount ||
-            candidateOption.partCount ||
-            candidateOption.selected
+            candidateItem.allCount ||
+            candidateItem.partCount ||
+            candidateItem.selected
           ) {
-            let relatedSelectedOption = find(
-              selectedOptions,
-              selectedOption => selectedOption.value === candidateOption.value
+            let relatedSelectedItem = find(
+              selectedItems,
+              selectedItem => selectedItem.value === candidateItem.value
             )
 
-            let newSelectedOption = omit(candidateOption, 'children')
-            if (this.hasChild(candidateOption)) {
-              newSelectedOption.children = walk(
-                candidateOption.children,
-                relatedSelectedOption && relatedSelectedOption.children
+            let newSelectedItem = omit(candidateItem, 'children')
+            if (this.hasChild(candidateItem)) {
+              newSelectedItem.children = walk(
+                candidateItem.children,
+                relatedSelectedItem && relatedSelectedItem.children
               )
 
-              // 如果右侧没有相同的 option ，说明当前这个 option 是新选中的。
-              // 对于新选中的 option ，要保持左侧的展开收起状态。
-              if (!relatedSelectedOption) {
-                let expanded = includes(this.candidateExpanded, newSelectedOption.value)
+              // 如果右侧没有相同的 item ，说明当前这个 item 是新选中的。
+              // 对于新选中的 item ，要保持左侧的展开收起状态。
+              if (!relatedSelectedItem) {
+                let expanded = includes(this.candidateExpanded, newSelectedItem.value)
                 if (expanded) {
-                  this.selectedExpanded.push(newSelectedOption.value)
+                  this.selectedExpanded.push(newSelectedItem.value)
                   uniq(this.selectedExpanded)
                 } else {
-                  remove(this.selectedExpanded, newSelectedOption.value)
+                  remove(this.selectedExpanded, newSelectedItem.value)
                 }
               }
             }
 
-            newSelectedOptions.push(newSelectedOption)
+            newSelectedItems.push(newSelectedItem)
           }
         })
-        return newSelectedOptions
+        return newSelectedItems
       }
-      return walk(this.candidateOptions, this.selectedOptions)
+      return walk(this.candidateItems, this.selectedItems)
     },
-    selectAllChildren (option, selected) {
-      if (this.hasChild(option)) {
-        option.children.forEach(child => {
+    selectAllChildren (item, selected) {
+      if (this.hasChild(item)) {
+        item.children.forEach(child => {
           if (this.hasChild(child)) {
-            this.setOptionCount(child, selected ? child.children.length : 0, 0)
+            this.setItemCount(child, selected ? child.children.length : 0, 0)
             this.selectAllChildren(child, selected)
           } else {
             this.setLeafSelected(child, selected)
@@ -361,46 +359,46 @@ export default {
       }
     },
     // 有可能用户传进来的 selected 没有在 datasource 里面，所以此处要处理一下
-    correct (options = this.candidateOptions) {
+    correct (items = this.candidateItems) {
       let allCount = 0
       let partCount = 0
-      options.forEach(option => {
-        if (this.hasChild(option)) {
-          let { allCount: all, partCount: part } = this.correct(option.children)
-          this.setOptionCount(option, all, part)
+      items.forEach(item => {
+        if (this.hasChild(item)) {
+          let { allCount: all, partCount: part } = this.correct(item.children)
+          this.setItemCount(item, all, part)
 
-          if (all === option.children.length) {
+          if (all === item.children.length) {
             allCount += 1
           } else if (all > 0 || part > 0) {
             partCount += 1
           }
         } else {
-          this.setLeafSelected(option, this.selected.some(v => v === option.value))
-          allCount += option.selected ? 1 : 0
+          this.setLeafSelected(item, this.selected.some(val => val === item.value))
+          allCount += item.selected ? 1 : 0
         }
 
-        this.$set(option, 'visuallySelected', this.isSelected(option))
+        this.$set(item, 'visuallySelected', this.isSelected(item))
       })
 
-      if (options === this.candidateOptions) {
+      if (items === this.candidateItems) {
         this.rootAllCount = allCount
         this.rootPartCount = partCount
       }
 
       return { allCount, partCount }
     },
-    findChain (options, value) {
-      let walk = (options, chain = []) => {
+    findChain (items, value) {
+      let walk = (items, chain = []) => {
         let currentChain = []
-        let result = options.some(option => {
-          if (option.value === value) {
-            currentChain.push(option)
+        let result = items.some(item => {
+          if (item.value === value) {
+            currentChain.push(item)
             return true
           }
 
-          if (this.hasChild(option)) {
-            currentChain.push(option)
-            return walk(option.children, currentChain)
+          if (this.hasChild(item)) {
+            currentChain.push(item)
+            return walk(item.children, currentChain)
           }
         })
 
@@ -413,7 +411,7 @@ export default {
       }
 
       let chain = []
-      walk(options, chain)
+      walk(items, chain)
       return chain
     }
   },
@@ -473,7 +471,7 @@ export default {
           CandidatePanel,
           {
             props: {
-              datasource: this.candidateOptions,
+              datasource: this.candidateItems,
               searchable: this.searchable,
               filter: this.filter,
               placeholder: this.candidatePlaceholder,
@@ -500,7 +498,7 @@ export default {
           SelectedPanel,
           {
             props: {
-              datasource: this.selectedOptions,
+              datasource: this.selectedItems,
               showMode: this.selectedShowMode,
               searchable: this.searchable,
               filter: this.filter,